+++ src/components/button/button.jsx
@@ -1,0 +1,6 @@
+//TODO: enhance later
+const Button = ({ children, type = 'button', ...props }) => {
+  return <button type={type} {...props} >{children}</button>
+}
+
+export default Button;
