+++ src/components/button/index.js
@@ -1,0 +1,4 @@
+import Button from './button';
+import './button.scss';
+
+export default Button;
