+++ src/components/header/header.jsx
@@ -1,36 +1,9 @@
-import React, { useRef } from 'react';
-import { useState, useEffect, useCallback } from "react";
-import { Link, NavLink, useSearchParams, createSearchParams } from "react-router-dom";
+import React from 'react';
+import { Link, NavLink } from "react-router-dom";
 import { useSelector } from 'react-redux';
-import { debounce } from "../../utils";
+import Search from '../search/search';
 
 const Header = () => {
-  const [searchParams, setSearchParams] = useSearchParams();
-  const searchQuery = searchParams.get('search');
-  const [search, setSearch] = useState(searchQuery || '');
-  const timeoutRef = useRef(null);
-
-  useEffect(() => {
-    setSearch(searchQuery || '');
-  }, [searchQuery]);
-
-  const getSearchResults = useCallback(query => {
-    if (query !== '') {
-      setSearchParams(createSearchParams({ search: query }), { replace: true })
-    } else {
-      setSearchParams()
-    }
-  }, [setSearchParams])
-
-  const debouncedSearchMovies = useCallback(value => {
-    debounce(getSearchResults, 500, timeoutRef)(value)
-  }, [getSearchResults])
-
-  const onSearch = ({ target: { value } }) => {
-    setSearch(value);
-    debouncedSearchMovies(value);
-  };
-
   const { starredMovies } = useSelector(state => state.starred);
 
   const renderStarredIcon = () => {
@@ -49,7 +22,6 @@ const Header = () => {
       <Link to="/" data-testid="home">
         <i className="bi bi-film" />
       </Link>
-
       <nav>
         <NavLink to="/starred" data-testid="nav-starred" className="nav-starred">
           {renderStarredIcon()}
@@ -58,19 +30,7 @@ const Header = () => {
           Watch Later
         </NavLink>
       </nav>
-
-      <div className="input-group rounded">
-        <input
-          type="search"
-          data-testid="search-movies"
-          onChange={onSearch}
-          className="form-control rounded"
-          placeholder="Search movies..."
-          aria-label="Search movies"
-          aria-describedby="search-addon"
-          value={search}
-        />
-      </div>
+      <Search />
     </header>
   );
 };
