+++ src/components/input/index.js
@@ -1,0 +1,4 @@
+import Input from './input'
+import './input.scss'
+
+export default Input
