+++ src/components/input/input.jsx
@@ -1,0 +1,7 @@
+
+//TODO: enhance later
+const Input = (props) => {
+  return <input {...props} />
+}
+
+export default Input
