+++ src/components/search/index.js
@@ -1,0 +1,4 @@
+import Search from './search'
+import './search.scss'
+
+export default Search
