+++ src/styles/mixins.scss
@@ -3,13 +3,13 @@ $tablet-width: 768px;
 $desktop-width: 1024px;
 
 @mixin mobile {
-  @media (min-width: #{$mobile-width}) {
+  @media (max-width: #{$tablet-width}) {
     @content;
   }
 }
 
 @mixin tablet {
-  @media (min-width: #{$tablet-width}) and (max-width: #{$desktop-width - 1px}) {
+  @media (min-width: #{$tablet-width}) and (max-width: #{$desktop-width}) {
     @content;
   }
 }
